feat: debug print 9p messages
This includes a command that can be used to display decoded 9p messages
using bytes from a file. The display format matches the logs from u9fs.

This also fixes a bug in the stat implementation. `Rstat` and `Twstat`
messages now use the `Stat` type rather than contain a bag of bytes. The
issue was that these messages had the size encoded twice, but we were
only decoding it once. I had failed to recognize this previously and
used the `ParsedStat` struct as an intermediate way to use the values
while keeping the stored value as a list of bytes. The fix is to
implement the double size prefix documented in the plan9 stat manpage.