fix: decode BitFlag::MAX
Previously we had a way to create a maximum value in the code, but
attempting to convert the numeric value to a flag would result in an
error due to truncation. Instead, when that error happens we check if
the maximum value has been specified. If so, we return as is.